base: lmp-device-register: add generic factory specific configs
Instead of relying on the factory to provide the device factory and api
via a bbappends, set them with a default value that can later be
replaced during CI automatically, removing the need for the extra
bbappend.

As we really only support aktualizr-lite at this stage, drop the extra
SOTA_CLIENT check and make it rdepend on aktualizr-lite by default.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>